DataBase Connection and Tables
I created the database inside of VS code and added the proper files. The names Program.cs will be responsible for getting all of the data from the tables in the csce361.sql file. Right now, there is not a connection set up between the csce361.sql and the Program.cs file to retrieve data.